steel_protocol

Module utils

Steel Protocol Utils

Utility functions and types for the protocol.

Structs

RawPacket

A raw packet.

StreamDecryptor

A stream that decrypts data.

StreamEncryptor

NOTE: This makes lots of small writes; make sure there is a buffer somewhere down the line

Enums

ConnectionProtocol

Describes the set of packets a connection understands at a given point.

PacketError

An error that can occur when handling packets.

Constants

MAX_PACKET_DATA_SIZE

The maximum size of a packet’s data.

MAX_PACKET_SIZE

The maximum size of a packet.

Type Aliases

Aes128Cfb8Dec

An AES-128 CFB-8 decryptor.

Aes128Cfb8Enc

An AES-128 CFB-8 encryptor.